refactored out SurveyCard, SurveyCardList, sorting works, filter wip
